requirements.txt
protobuf==5.26.1
pydantic==2.7.0
PyYAML==6.0.1
sgp4==2.23
skyfield==1.48
paho-mqtt==2.0.0
pandas==2.2.2
